% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BCR}
\alias{BCR}
\title{balanced classification rate (BCR)
The BCR computes the classification rate for each class and reports the arithmetic average of those rates over all classes}
\usage{
BCR(cm)
}
\arguments{
\item{cm}{a confusion matrix}
}
\value{
the bcr
}
\description{
balanced classification rate (BCR)
The BCR computes the classification rate for each class and reports the arithmetic average of those rates over all classes
}
