% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{benchmark}
\alias{benchmark}
\title{run the given function wrapped around two calls of Sys.time() and returns the execution time delta and the function result}
\usage{
benchmark(f, args)
}
\arguments{
\item{f}{The function on which to perform the benchmark}

\item{args}{the function arguments}
}
\value{
A list containing tdelta, the execution time and fresult, the function's return.
}
\description{
run the given function wrapped around two calls of Sys.time() and returns the execution time delta and the function result
}
\examples{
benchmark(mean, list(c(1,2,3)))
}
