% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pkg_score.R
\name{calculate_pkg_score}
\alias{calculate_pkg_score}
\title{Convert the text risk metric answers to numeric values and calculate a score}
\usage{
calculate_pkg_score(pkgs = NULL)
}
\arguments{
\item{pkgs}{a list of package objects}
}
\description{
The score is the mean of the individual scores for each question.
}
