% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{change_class}
\alias{change_class}
\title{Randomly modify the class label of a \code{proportion} of observations from \code{data}}
\usage{
change_class(data, class = "class", proportion = 0.1)
}
\arguments{
\item{data}{The original dataset}

\item{class}{The name of the class variable. class by default}

\item{percentage}{The percentage of affected labels}
}
\value{
A modified copy of \code{data} containing \code{proportion} modified labels from \code{class} and vector indicating which indices have been modified
}
\description{
Randomly modify the class label of a \code{proportion} of observations from \code{data}
}
\examples{
change_class(data)
change_class(data, class="y")
change_class(data, proportion=.5)
}
