% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_session.R
\name{check_session}
\alias{check_session}
\alias{print.sctovalidity}
\title{Check a sessions loaded packages against the validated packages list}
\usage{
check_session(attached_only = TRUE, approved_only = TRUE)

\method{print}{sctovalidity}(x, ...)
}
\arguments{
\item{attached_only}{logical, should all loaded packages be shown (FALSE; default), or only those that are attached (TRUE)}

\item{approved_only}{logical, should only validated (approved) packages be shown (TRUE; default), or also those awaiting approval (FALSE)}

\item{x}{output from \code{check_session}}

\item{...}{options passed to methods}
}
\description{
Check a sessions loaded packages against the validated packages list
}
\examples{
\dontrun{
check_session()
}
}
