% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{choose_color}
\alias{choose_color}
\title{A function that returns a vector of colors going from start to end}
\usage{
choose_color(n = 10, start = "green", end = "red")
}
\arguments{
\item{n}{number of points}

\item{start}{a color}

\item{end}{a color}
}
\description{
A function that returns a vector of colors going from start to end
}
