% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_pack.R
\name{contri_file_create}
\alias{contri_file_create}
\title{Job/State/MSA contribution}
\usage{
contri_file_create(data, group_by_elements = NA, eda_var, year,
  contri_of)
}
\arguments{
\item{data:}{The data frame on which the basic EDA needs to be done.}

\item{group_by_element:}{the variables on which the stats var will be calculated. For example: for country, group_by_var="", for state occupation granularity, it will be c(state_id).
Note the table should be in same granularity level as the group_by_elements.}

\item{eda_var:}{The attribute name on which eda needs to be performed}

\item{year:}{Attribute that has year.}

\item{contri_of:}{This variable tells about the attribute for which the contribution needs to calculated.}
}
\description{
The function is used to calculate contribution of any attribute based on its supply.
}
