% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_downloads_to_score.R
\name{convert_downloads_to_score}
\alias{convert_downloads_to_score}
\title{Convert number of downloads to a score
The score is based on the approach by the riskmetric package. It is based on a
simplification of the logistic curve.}
\usage{
convert_downloads_to_score(downloads)
}
\arguments{
\item{downloads}{number of downloads}
}
\description{
See \href{https://pharmar.github.io/riskmetric/reference/metric_score.pkg_metric_downloads_1yr.html}{riskmetric documentation} for further details.
}
\seealso{
\href{https://pharmar.github.io/riskmetric/reference/metric_score.pkg_metric_downloads_1yr.html}{riskmetric documentation}
}
\keyword{internal}
