% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correctely_detected}
\alias{correctely_detected}
\title{returns the percentage of correctly detected outliers given that we consider a given percentage of the dataset to be outlier}
\usage{
correctely_detected(scores, percentage, y)
}
\arguments{
\item{scores}{a vector of outlier scores}

\item{percentage}{the percentage of dataset to be considered as outlier}

\item{y}{the real outliers}
}
\value{
the percentage of correctly detected outliers
}
\description{
returns the percentage of correctly detected outliers given that we consider a given percentage of the dataset to be outlier
}
