% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset1.R
\docType{data}
\name{dataset1}
\alias{dataset1}
\title{Anonymized example of data set provided by DNAlytics}
\format{An object of class \code{data.frame} with 48 rows and 2589 columns.}
\usage{
data(dataset1)
}
\description{
Anonymized example of data set provided by DNAlytics
}
\examples{
data(dataset1)
class <- dataset1$class
}
\keyword{datasets}
