% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_mapper.R
\name{domain_mapper}
\alias{domain_mapper}
\title{Map a domain from one range to another}
\usage{
domain_mapper(
  value,
  min_original = 0,
  max_original = 1,
  min_range = 0,
  max_range = 1
)
}
\arguments{
\item{value}{value to map from}

\item{min_original}{minimum of the original domain}

\item{max_original}{maximum of the original domain}

\item{min_range}{minimum of the desired domain}

\item{max_range}{maximum of the desired domain}
}
\value{
values mapped to the new domain, typically a vector or single value.
}
\description{
Map a domain from one range to another
}
\examples{
domain_mapper(0.5, 0, 1, 0, 100)
}
\keyword{internal}
