% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_pack.R
\name{eda_file_create}
\alias{eda_file_create}
\title{Create EDA file}
\usage{
eda_file_create(data, group_by_elements, eda_var, year)
}
\arguments{
\item{data:}{The data frame on which the basic EDA needs to be done.}

\item{group_by_element:}{the variables on which the stats var will be calculated. For example: for country, group_by_var="occupation_code", for state occupation granularity, it will be c(state_id, occupation_code).
Note the table should be in same granularity level as the group_by_elements.}

\item{eda_var:}{The attribute name on which eda needs to be performed}

\item{year:}{Attribute that has year.}
}
\description{
The function is used to calculate the CV of the supply and the growth rates.
}
