% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_matlab}
\alias{export_matlab}
\title{export matlab data to R file}
\usage{
export_matlab(filename, origin = "data/dataPublic/",
  destination = "data/converted/")
}
\arguments{
\item{filename}{the name of the matlab file}

\item{origin}{the folder of the file}

\item{destination}{the destination folder}
}
\description{
export matlab data to R file
}
