% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elements_test.R
\name{extract_elements_test}
\alias{extract_elements_test}
\title{extract the useful text elements from a test issue}
\usage{
extract_elements_test(issue)
}
\description{
extract the useful text elements from a test issue
}
\keyword{internal}
