% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{gen_multi_class}
\alias{gen_multi_class}
\title{Generates a dataset composed of normal correleted features with two classes.}
\usage{
gen_multi_class(n, normal.q = 0.9, outliers.q = 0.01, mu_a = rep(0,
  nrow(cov_a)), mu_b = rep(0, nrow(cov_b)), mu_global = NULL,
  cov_a = diag(5), cov_b = diag(5), cov_global = diag(5))
}
\arguments{
\item{n}{the number of samples}

\item{normal.q}{the percentile to consider as normal samples}

\item{outliers.q}{the percentile to consider as outliers}

\item{mu_a}{the means of the class a}

\item{mu_b}{the means of the class b}

\item{mu_global}{the means of the extra non significant variables. can be NULL}

\item{cov_a}{covariance matrix of class a}

\item{cov_b}{covariance matrix of class b}

\item{cov_global}{covariance matrix of extra non significant variables. can be NULL}
}
\value{
list containing two sets, normal and outliers
}
\description{
Generates a dataset composed of normal correleted features with two classes.
}
