% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{gen_normal}
\alias{gen_normal}
\title{Generates a dataset composed of normal correleted features with one classes.}
\usage{
gen_normal(n = 10000, normal.q = 0.9, outliers.q = 0.01, d = 50,
  mu = rep(0, d), sigma = diag(d))
}
\arguments{
\item{n}{the number of samples}

\item{normal.q}{the percentile to consider as normal samples}

\item{outliers.q}{the percentile to consider as outliers}

\item{mu}{the means}

\item{sigma}{covariance matrix}
}
\value{
list containing two sets, normal and outliers
}
\description{
Generates a dataset composed of normal correleted features with one classes.
}
