% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pkg_table.R
\name{gen_pkg_table}
\alias{gen_pkg_table}
\title{Generate a table with the package validation statistics}
\usage{
gen_pkg_table(pkgs = NULL, ...)
}
\arguments{
\item{pkgs}{NULL or a list of issues from \code{get_issues}}
}
\description{
This is intended for use by a github action to generate a table for the website
}
