% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_12month_downloads.R
\name{get_12month_downloads}
\alias{get_12month_downloads}
\title{Get number of downloads in the last 12 months}
\usage{
get_12month_downloads(package, from = Sys.Date() - 365, to = Sys.Date())
}
\arguments{
\item{package}{package name}

\item{from}{start date}

\item{to}{end date}
}
\value{
integer number of downloads
}
\description{
Note: this is only available for CRAN or bioconductor packages
}
\examples{
get_12month_downloads("ggplot2")
get_12month_downloads("secuTrialR")

}
