% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R, R/close_issue.R, R/get_comments.R,
%   R/get_issue.R, R/get_issues.R, R/get_test_reports.R, R/post_comment.R,
%   R/post_issue.R, R/remove_label.R, R/update_issue.R
\name{add_label}
\alias{add_label}
\alias{close_issue}
\alias{get_comments}
\alias{get_issue}
\alias{get_issues}
\alias{get_test_reports}
\alias{post_comment}
\alias{post_issue}
\alias{remove_label}
\alias{update_issue}
\title{Interact with the GitHub API to get or set issue information}
\usage{
add_label(issue, label, repo = sctoreports())

close_issue(issue, repo = sctoreports())

get_comments(issue, repo = sctoreports())

get_issue(issue, repo = sctoreports())

get_issues(repo = sctoreports())

get_test_reports(repo = sctoreports(), approved_only = FALSE)

post_comment(issue, comment, repo = sctoreports())

post_issue(body, title, repo = sctoreports())

remove_label(issue, label, repo = sctoreports())

update_issue(issue, ..., repo = sctoreports())
}
\arguments{
\item{issue}{issue number to update}

\item{label}{label to remove}

\item{repo}{repository name

... could be e.g. body, title, labels (should be a list), state (open or closed)
see https://docs.github.com/en/rest/issues/issues?apiVersion=2022-11-28#update-an-issue}

\item{approved_only}{only return approved packages (logical)}

\item{comment}{the comment to be made}

\item{body}{issue body}

\item{title}{issue title}

\item{...}{arguments passed to gh}
}
\value{
a list of comments

dataframe
}
\description{
These functions provide various ways to interact with github issues.

Open/update an issue:

\code{post_issue} opens a new issue in a repository, with a defined body and title.

\code{update_issue} updates an existing issue in a repository (could also be used
instead of \code{add_labels} or \code{close_issue}).

Getting issues:

\code{get_issue} downloads a single issue from a repository.

\code{get_issues} downloads all issues from a repository.

\code{get_test_reports} downloads all issues from a repository, filtering those with
test labels, and converts the issue to a dataframe.

Labels:

\code{add_label} adds a particular label to an issue.

\code{remove_label} removes a particular label from an issue.

Comments:

\code{get_comments} downloads all comments from an issue.

\code{post_comment} posts a comment to an issue.

Closing an issue:

\code{close_issue} closes an issue.

These functions are primarily for the use in github actions rather than for
the standard user.

See https://docs.github.com/en/rest/issues/issues for full information.
}
\examples{
# remove_label(issue = 1, label = "test")

# post_comment(issue = 1, comment = "This is a test comment")
# get_issue(21)
# post_comment(issue = 1, comment = "This is a test comment")
# post_issue("Some body text", title = "Issue title")
# remove_label(issue = 1, label = "test")

# post_issue("Some body text", title = "Issue title")
}
