% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_deps.R
\name{get_n_deps}
\alias{get_n_deps}
\title{Get the number of dependencies for a package}
\usage{
get_n_deps(package, ...)
}
\arguments{
\item{package}{package name}

\item{...}{additional arguments passed to
\code{\link{tools::package_dependencies}}. See details for more information.}
}
\value{
integer: the number of dependencies
}
\description{
Get the number of dependencies for a package
}
\details{
The \code{recursive} and \code{reverse} arguments are perhaps of most interest.
\code{recursive} details the number of dependencies of the package, including
those that are dependencies of dependencies. \code{reverse} details the number
of packages that depend on the package, which may be an indicator of a high
quality package.
}
\examples{
get_n_deps("nlme")
get_n_deps("validation")
get_n_deps("dplyr", reverse = TRUE)

}
\seealso{
\code{\link{tools::package_dependencies}}
}
