% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_source.R
\name{get_pkg_source}
\alias{get_pkg_source}
\title{Derive the location of a package's source code}
\usage{
get_pkg_source(pkg, ...)
}
\arguments{
\item{pkg}{package to check the source of}

\item{...}{parameters passed to other methods}
}
\description{
This function tries to determine where a package was installed from, e.g. CRAN,
GitHub, r-universe, etc. If the package comes from a git repository, it also
attempts to extract the repository URL and commit hash.
}
\details{
If the package was installed from a r-universe, the function will return the
git repository URL and commit hash of the package, rather than the r-universe
URL. E.g. this package is installed from the CTU-Bern r-universe, but the function
returns the URL of the GitHub repository
(\code{https://github.com/SwissClinicalTrialOrganisation/validation}) and the
commit hash.
}
