% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{get_pkgs}
\alias{get_pkgs}
\title{get packages from github, or filter a list of issues for packages}
\usage{
get_pkgs(issues = NULL, ...)
}
\description{
get packages from github, or filter a list of issues for packages
}
\keyword{internal}
