% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_date.R
\name{get_release_date}
\alias{get_release_date}
\title{Get the release date of a CRAN package}
\usage{
get_release_date(pkg)
}
\arguments{
\item{pkg}{a package name}
}
\value{
character string containing either the date or a message to check elsewhere
}
\description{
CRAN packages have a release date associated with them. This function downloads
that date and returns it as a date object.
}
\details{
If the package does not exist on CRAN, the function asks the user to search
elsewhere for the information.
}
\examples{

get_release_date("dplyr")

}
