% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_sha.R
\name{get_repo_sha}
\alias{get_repo_sha}
\title{get the most recent commit SHA for a repo}
\usage{
get_repo_sha(repo = sctotests(), branch = "main")
}
\arguments{
\item{repo}{github repository to download from}

\item{branch}{branch to get the SHA from}
}
\description{
get the most recent commit SHA for a repo
}
\keyword{internal}
