% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Download datasets from the SCTO validation tests repository}
\usage{
get_test_data(dataset, repo = sctotests(), dir = getwd())
}
\arguments{
\item{dataset}{The filename of the dataset to download}

\item{repo}{the repository to download from}

\item{dir}{the directory to save the dataset to (purely for testing purposes
- this option should not be used in everyday use)}
}
\value{
TRUE if the dataset is downloaded successfully
}
\description{
Tests sometimes require specific datasets. Where a built-in dataset is insufficient,
a dataset that exists within the SCTO validation tests repository can be downloaded.
}
\details{
This function can be used in the setup-pkg.R file so that the datasets are available
to the tests.


The function will download the dataset from the SCTO validation tests repository
to the working directory, which testthat sets to the directory containing the
test files. Therefore, the dataset can be downloaded and saved and used without
referring to any other directory. The setup-pkg.R file might then contain:

\preformatted{
...
library(testthat)

get_test_data("mtcars.csv")
dat <- read.csv("mtcars.csv")

withr::defer({})
}
}
\examples{
# download the file to a temporary directory
tempdir <- tempdir()
get_test_data("mtcars.csv", dir = tempdir)
dat <- read.csv(file.path(tempdir, "mtcars.csv"))
unlink(tempdir)

\dontrun{
# within the setup-pkg.R file
get_test_data("mtcars.csv")
# then load the data using an appropriate function
dat <- read.csv("mtcars.csv")
}

}
