% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tested_pkgs.R
\name{get_tested_pkgs}
\alias{get_tested_pkgs}
\title{Get the list of packages that have tests in a repository}
\usage{
get_tested_pkgs(repo = sctotests())
}
\arguments{
\item{repo}{repository to check for tests}
}
\value{
character vector of package (directory) names
}
\description{
Tests are assumed to be in a `tests` directory in the repository.
}
\examples{
get_tested_pkgs()
}
