% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests.R
\name{get_tests}
\alias{get_tests}
\title{download test files from the validation_tests repo}
\usage{
get_tests(pkg, dir = tempdir(), repo = sctotests())
}
\arguments{
\item{pkg}{package name as a string}

\item{dir}{directory to download to}

\item{repo}{github repository to download from}
}
\description{
download test files from the validation_tests repo
}
\keyword{internal}
