% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_pkgs.R
\name{get_valid_pkgs}
\alias{get_valid_pkgs}
\title{Get the list of validated packages from GitHub}
\usage{
get_valid_pkgs(approved_only = TRUE)
}
\arguments{
\item{approved_only}{only return approved packages (logical)}
}
\value{
dataframe including package, version, etc.
}
\description{
This function downloads issues from github and extracts relevant information from them
}
\examples{
# only approved packages
get_valid_pkgs()
# all packages, regardless of status
get_valid_pkgs(FALSE)
}
