% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{graph}
\alias{graph}
\title{Plot a graph with outliers in red}
\usage{
graph(lor, lensemble, y = NULL, n = 10, percent = 0.2)
}
\arguments{
\item{lor}{list of result of all functions}

\item{lensemle}{vector of result of the ensemble function}

\item{max_n}{maximum number of points draw}
}
\description{
Plot a graph with outliers in red
}
