% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{introduce_perturbation}
\alias{introduce_perturbation}
\title{Return a dataset with the specified proportion of the sample corrupted by the function \code{fun}.}
\usage{
introduce_perturbation(data, proportion, fun)
}
\arguments{
\item{data}{The original data set}

\item{proportion}{The proportion of that data needs to be corrupted}

\item{fun}{the corrupting function. takes a data set and a vector of boolean indicating which data is corrupted as argument and return a corrupted data set}
}
\value{
a list containing: data the corrupted data set , is_modified the vector indicating which observation has been corrupted
}
\description{
Return a dataset with the specified proportion of the sample corrupted by the function \code{fun}.
}
\examples{
introduce_perturbation(dataset1, .1, fun(data, is_modified){
data[is_modified, ]$f1  <- 0
data
} )
}
