% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{introduce_value_increased_outliers}
\alias{introduce_value_increased_outliers}
\title{Return a dataset with the specified proportion of the sample corrupted with the specified \code{feature_proportion} modified with random noise between noise_min and noise_max}
\usage{
introduce_value_increased_outliers(data, sample_proportion = 0.1,
  feature_proportion = 0.1, noise_min = 0.2, noise_max = 0.7)
}
\arguments{
\item{data}{The original data set}

\item{sample_proportion}{The proportion of that data needs to be corrupted}

\item{feature_proportion}{the proportion of the features that are affected}

\item{noise_min}{the minimum percentage of noise added to the features as a percentage of itself. the new value will be value + value*noise (with noise, a value between noise_min and noise_max)}

\item{noise_max}{the maximal percentage of noise added to the features as a percentage of itself. the new value will be value + value*noise (with noise, a value between noise_min and noise_max)}
}
\value{
a list containing: data the corrupted data set, is_modified the vector indicating which observation has been corrupted
}
\description{
Return a dataset with the specified proportion of the sample corrupted with the specified \code{feature_proportion} modified with random noise between noise_min and noise_max
}
\examples{
introduce_value_increased_outliers(dataset1, proportion=.1, distance_min = .0, distance_max=.2, affected_features=.20)
}
