% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_}
\alias{is_}
\alias{is_package}
\alias{is_test}
\alias{is_approved}
\alias{is_triage}
\alias{get_labels}
\title{Check whether an issue has a particular label}
\usage{
is_(issue, what)

is_package(issues)

is_test(issues)

is_approved(issues)

is_triage(issues)

get_labels(issue)
}
\arguments{
\item{issue}{a list of issues}

\item{what}{the label to check for}
}
\description{
These functions provide a way to check whether an issue has a particular label.

\code{is_} is the main function. The others are wrappers around this one.

\code{is_package} checks for a package label.

\code{is_test} checks for a test label.

\code{is_approved} checks for an approved label.

\code{is_triage} checks for a triage label.

\code{get_labels} gets all labels associated with an issue.
}
\examples{
# issues <- get_issues()
# is_package(issues)

# issue <- get_issue(21)
# is_package(list(issue))
}
