% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg_table.R
\name{load_pkg_table}
\alias{load_pkg_table}
\title{Load the flat file of validated packages from github}
\usage{
load_pkg_table()
}
\description{
Github contains a flat file of the package validation results. This function
retrieves that file.
}
