% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{pca_display}
\alias{pca_display}
\title{generate pdf containing a pca for specified dataset for each method in v_of_techniques}
\usage{
pca_display(original_dataset_name, dataset_name, v_of_techniques,
  path = "data/result/", origin = "data/converted/", pca_method = prcomp,
  preprocess_method = outliers::preprocessing)
}
\arguments{
\item{original_dataset_name}{original name of the dataset}

\item{dataset_name}{name of the dataset}

\item{v_of_techniques}{vector containing the different detection techniques}

\item{path}{The path containing the results fom the different detection methods}

\item{origin}{The path containing the dataset}

\item{pca_method}{The method to use to compute the pca}
}
\description{
generate pdf containing a pca for specified dataset for each method in v_of_techniques
}
