% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot a boxplot with detected outliers in red. return the index of the points outside of the wiskers}
\usage{
plot_boxplot(scores, outliers = NULL, nsd = 5, threshold = mean(scores) +
  nsd * sd(scores), show_names = TRUE, title = "")
}
\arguments{
\item{scores}{the scores to plot}

\item{nsd}{the number of standard deviation to be considered as outliers}

\item{threshold}{the max score for normal point}
}
\description{
Plot a boxplot with detected outliers in red. return the index of the points outside of the wiskers
}
