% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot a pca with detected outliers in red}
\usage{
plot_pca(data, scores, pca = prcomp(data), y = rep(FALSE, length(scores)),
  out_shape = 17, normal_shape = 19, nsd = 4, title = "")
}
\arguments{
\item{data}{normalized data}

\item{scores}{the scores to plot}

\item{pca}{the pca projection}

\item{y}{the outliers}

\item{out_shape}{the shape to use to represent outliers}

\item{normal_shape}{the shape to use to represent normal points}

\item{nsd}{the number of standard deviation to be considered as outliers}
}
\description{
Plot a pca with detected outliers in red
}
