% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rda_to_list}
\alias{rda_to_list}
\title{open a file and return the labels (y)}
\usage{
rda_to_list(dataset_name, v_of_techniques, path = "data/result/")
}
\arguments{
\item{dataset_name}{name of the files (without.RDA)}

\item{v_of_techniques}{list of techniques uses}

\item{path}{path to the document}

\item{files}{name of the files (without.RDA)}

\item{args}{the function arguments}

\item{path}{path to the document}
}
\value{
a vector y

a list of results
}
\description{
open a file and return the labels (y)
}
