% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_repos.R
\name{sctoreports}
\alias{sctoreports}
\alias{sctotests}
\alias{sctopkg}
\title{SCTO repositories}
\usage{
sctoreports()

sctotests()

sctopkg()
}
\value{
the repository name (validation)
}
\description{
shortcuts for the SCTO package validation repositories. The validation platform
consists of 3 repositories: a repository for reporting the validations
(pkg_validation), one which contains tests for specific packages
(validation_tests), and one for this package itself (validation). These three
functions provide the names of the three repositories.
}
