% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{resample}
\alias{resample}
\title{helper method to resample a data set containing 1 classe.}
\usage{
resample(data, nnormal = 100, noutliers = ceiling(5 * nnormal/100),
  d = ncol(data$normal), replace = FALSE)
}
\arguments{
\item{data}{the dataset}

\item{nnormal}{number of normal points}

\item{noutliers}{number of outliers}

\item{replace}{TRUE if the same sample can be picked multiple times}

\item{normal.q}{the percentile to consider as normal samples}
}
\value{
list containing: x the new dataset, y a boolean vector indicating which sample is an outlier
}
\description{
helper method to resample a data set containing 1 classe.
}
