% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{resample_multi}
\alias{resample_multi}
\title{helper method to resample a data set containing 2 classes.}
\usage{
resample_multi(data, nnormal = 100, noutliers = ceiling(5 * nnormal/100),
  d = ncol(data$normal), replace = FALSE, balance = 0.5)
}
\arguments{
\item{data}{the dataset}

\item{nnormal}{number of normal points}

\item{noutliers}{number of outliers}

\item{replace}{TRUE if the same sample can be picked multiple times}

\item{balance}{specify the class balance.}

\item{normal.q}{the percentile to consider as normal samples}
}
\value{
list containing: x the new dataset ( with an extra feature for the class), y a boolean vector indicating which sample is an outlier
}
\description{
helper method to resample a data set containing 2 classes.
}
