% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_pack.R
\name{save_file}
\alias{save_file}
\title{Save excel function}
\usage{
save_file(wb, geog)
}
\arguments{
\item{eb:}{workbook object}

\item{geog:}{The geography for which the file is saved}
}
\description{
A function to save the excel workbook onto the local system
}
