% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{split_normal_outliers}
\alias{split_normal_outliers}
\title{helper method to separate normal points, outliers and noise in a normal correleted dataset}
\usage{
split_normal_outliers(data, normal.q, outliers.q, mu, sigma)
}
\arguments{
\item{data}{the dataset}

\item{normal.q}{the percentile to consider as normal samples}

\item{outliers.q}{the percentile to consider as outliers}

\item{mu}{the means}

\item{sigma}{covariance matrix}
}
\value{
list containing three sets, normal noise and  outliers
}
\description{
helper method to separate normal points, outliers and noise in a normal correleted dataset
}
