% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_skeleton.R
\name{test_skeleton}
\alias{test_skeleton}
\title{Add a new package skeleton}
\usage{
test_skeleton(pkg, funs, dir = getwd())
}
\arguments{
\item{pkg}{package to be tested}

\item{fun}{function to be tested}
}
\value{
a set of files in your working directory, which can be edited and
  uploaded to github
}
\description{
This function will create the directory if it does not exist, and add a `info.txt` file with the package name, and a `setup-pkg.R` file with the necessary setup code.
}
