% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_to_text.R
\name{test_to_text}
\alias{test_to_text}
\title{Convert a test result object to a text string for posting to GitHub}
\usage{
test_to_text(x)
}
\arguments{
\item{x}{test result object}
}
\value{
a list of character strings
}
\description{
Convert a test result object to a text string for posting to GitHub
}
\examples{
# texts <- test_to_text(x)
# issue <- post_issue(texts$issue_body, texts$issue_title)
# map(texts$issue_tags, ~ add_label(issue$number, .x))
}
