% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pkg_issue.R, R/validate_test_issue.R
\name{validate_pkg_issue}
\alias{validate_pkg_issue}
\alias{validate_test_issue}
\title{Validate the inputs for package or function tests}
\usage{
validate_pkg_issue(score)

validate_test_issue(issue)
}
\arguments{
\item{score}{the output from calculate_pkg_score for a single package}

\item{issue}{a single function test issue}
}
\value{
list wit two elements: score_ok (logical) and message (character)

list wit two elements: ok (logical) and message (character)
}
\description{
These functions validate the inputs from package or function tests.

Both functions return a list with two elements - an indicator of whether the
validation passed without any errors and a character indicating what the error(s)
was/were. The character string is formated as if it were a comment to a GitHub
issue.

The functions are not intended to be used directly by the user, but are called
by a github action with the results posted as a comment to github.

`validate_pkg_issue` checks the values from a package issue.

`validate_test_issue` checks the values from a function test issue.
}
