% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elements_pkg.R
\name{extract_elements_pkg}
\alias{extract_elements_pkg}
\title{extract the useful text elements from an issue}
\usage{
extract_elements_pkg(issue)
}
\description{
extract the useful text elements from an issue
}
\keyword{internal}
