% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pkg_table.R, R/gen_tests_table.R
\name{gen_pkg_table}
\alias{gen_pkg_table}
\alias{gen_tests_table}
\title{Convert issues from to tables}
\usage{
gen_pkg_table(pkgs = NULL, ...)

gen_tests_table(tests = NULL, ...)
}
\arguments{
\item{pkgs}{NULL or a list of issues from \code{get_issues}}

\item{...}{options passed to \code{get_issues}}

\item{tests}{NULL or a list of issues from \code{get_issues}}
}
\description{
These functions are intended for use by a github action to generate a table for the website

This is intended for use by a github action to generate a table for the website
}
\section{Functions}{
\itemize{
\item \code{gen_tests_table()}: Generate a table of function tests

}}
