% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_to_df.R
\name{issues_to_df}
\alias{issues_to_df}
\title{issues are a list of items. this function converts it to a dataframe}
\usage{
issues_to_df(issues, fun = extract_elements_pkg)
}
\description{
issues are a list of items. this function converts it to a dataframe
}
\keyword{internal}
