% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_evidence_tab.R
\name{parse_evidence_tab}
\alias{parse_evidence_tab}
\title{Convert test evidence table back to a data frame}
\usage{
parse_evidence_tab(tab)
}
\arguments{
\item{tab}{a string containing the markdown table}
}
\value{
a data frame
}
\description{
In order to print nicely in GitHub, tables are formatted using markdown syntax.
}
\details{
This function can read that syntax and convert it back to a normal data frame.
}
\examples{
tab <- c("|file      |context |test      | nb| passed|skipped |error | warning|",
         "|:---------|:-------|:---------|--:|------:|:-------|:-----|-------:|",
         "|test-lm.R |lm      |message 1 |  8|      4|FALSE   |FALSE |       4|",
         "|test-lm.R |lm      |message 2 |  3|      3|FALSE   |FALSE |       0|")
parse_evidence_tab(tab)
}
