% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Run all tests for a given package}
\usage{
test(pkg, download = TRUE, cleanup = FALSE, dir = getwd(), repo = sctotests())
}
\arguments{
\item{pkg}{package name as a string}

\item{download}{whether to download a fresh set of tests}

\item{cleanup}{whether to delete testing folder}

\item{dir}{directory to download tests to. This should have \code{pkg} at the end. If not present, it will be created.}

\item{repo}{storing the tests}
}
\value{
Object of class \code{validate_result}
}
\description{
This function runs all predefined tests for a given package.
}
\details{
The print method supports the creation of a new issue in the package's
repository by presenting the information in a format that can be copied and
pasted in the issue template.
}
\examples{
if(FALSE) test("stats")
}
