% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pkg_table.R, R/update_tests_table.R
\name{update_pkg_table}
\alias{update_pkg_table}
\alias{update_tests_table}
\title{Append new package data to existing package table}
\usage{
update_pkg_table(...)

update_tests_table(...)
}
\arguments{
\item{...}{options passed to gen_pkg_table e.g. a list of issues from get_pkgs}
}
\description{
Where a package version is already in the table, the new data will replace the old data.

Where a package version is already in the table, the new data will replace the old data.
}
\section{Functions}{
\itemize{
\item \code{update_tests_table()}: Append new test data to existing tests table

}}
